+++ core/src/test/java/com/questdb/cairo/TableUtilsTest.java
@@ -46,7 +46,7 @@ public class TableUtilsTest {
         try (CompositePath path = new CompositePath()) {
             Assert.assertEquals(0, TableUtils.exists(root, metadata.getName()));
 
-            try (ReadOnlyMemory mem = new ReadOnlyMemory(path.of(root).concat(metadata.getName()).concat("_meta").$(), (int) Files.PAGE_SIZE, 0)) {
+            try (ReadOnlyMemory mem = new ReadOnlyMemory(path.of(root).concat(metadata.getName()).concat("_meta").$(), Files.PAGE_SIZE, 0)) {
                 long p = 0;
                 Assert.assertEquals(metadata.getColumnCount(), mem.getInt(p));
                 p += 4;
